Tweaks to solid pattern.
I think it's working, but it may not be the correct way to do this. I'm
going to try a new branch, in which instead of scanning from every
possible camera point, I just start from every point in the image
(vertex, etc.) and trace outward horizontally and then see if that ray
hits anything.