Visibility is now checked by rotating the viewpoint about the origin.
Overall, I'm not satisfied with the results this produces, and I think
my simulator may not be giving me accurate output. Positioning the
camera correctly is also annoying and may be difficult for people who
aren't as familiar with what the hell is going on as I am. I think my
next step should be to do these things:

TODO:
1) Revert to the initial system in which the z value of each point is
its depth
2) Change the PDFPrinter to a format I can trace with a compass (or add
another printout or something)
3) Produce a simple opacity demo that I can make by hand so I can
actually test this system properly